move platform subclasses to engines and runners
