+++ src/main/scala/io/viash/engines/docker/package.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.viash.platforms
+package io.viash.engines
 
 import io.circe.{Decoder, Encoder, Json}
 import io.circe.generic.extras.semiauto.{deriveConfiguredDecoder, deriveConfiguredEncoder}
@@ -35,20 +35,6 @@ package object docker {
           case "auto" | "automatic" => Automatic
         }
       )
-  }
-
-  // encoder and decoder for setup strategy
-  implicit val encodeSetupStrategy: Encoder[DockerSetupStrategy] = Encoder.instance {
-    dss => Json.fromString(dss.id.toLowerCase)
-  }
-  implicit val decodeSetupStrategy: Decoder[DockerSetupStrategy] = Decoder.instance {
-    cursor =>
-      cursor.value.as[String].map { s =>
-        val id = s.toLowerCase.replaceAll("_", "")
-        DockerSetupStrategy.map.applyOrElse(id,
-          (key: String) => throw new Exception(s"Docker Setup Strategy '$key' not found.")
-        )
-      }
   }
 
 }
