+++ src/main/scala/io/viash/platforms/DockerPlatform.scala
@@ -24,14 +24,15 @@ import io.viash.config.Config
 import io.viash.functionality._
 import io.viash.functionality.arguments._
 import io.viash.functionality.resources._
-import io.viash.platforms.requirements._
 import io.viash.helpers.{Bash, Docker}
 import io.viash.wrapper.{BashWrapper, BashWrapperMods}
-import io.viash.platforms.docker._
 import io.viash.helpers.data_structures._
 import io.viash.config.Info
 import io.viash.schemas._
 import io.viash.helpers.Escaper
+import io.viash.runners.executable.{IfNeedBePullElseCachedBuild, DockerSetupStrategy}
+import io.viash.engines.requirements.Requirements
+import io.viash.engines.docker.{DockerResolveVolume, Automatic}
 
 
 @description(
