+++ src/main/scala/io/viash/platforms/NativePlatform.scala
@@ -20,7 +20,6 @@ package io.viash.platforms
 import io.viash.config.Config
 import io.viash.functionality.Functionality
 import io.viash.functionality.resources._
-import io.viash.platforms.requirements._
 import io.viash.wrapper.BashWrapper
 import io.viash.schemas._
