+++ src/main/scala/io/viash/runners/NextflowRunner.scala
@@ -23,7 +23,6 @@ import io.viash.functionality.resources._
 import io.viash.functionality.arguments._
 import io.viash.helpers.{Docker, Bash, DockerImageInfo, Helper}
 import io.viash.helpers.circe._
-import io.viash.platforms.nextflow._
 import io.circe.syntax._
 import io.circe.{Printer => JsonPrinter, Json, JsonObject}
 import shapeless.syntax.singleton
@@ -31,6 +30,7 @@ import io.viash.schemas._
 import io.viash.helpers.Escaper
 import io.viash.runners.{Runner, RunnerResources}
 import io.viash.engines.DockerEngine
+import io.viash.runners.nextflow._
 
 @description(
   """Run a Viash component on a Nextflow backend engine.
