+++ src/main/scala/io/viash/runners/nextflow/NextflowDirectives.scala
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.viash.platforms.nextflow
+package io.viash.runners.nextflow
 
 import io.viash.helpers.data_structures._
 import io.viash.schemas._
